adding casting spell script and changed to the menu system - NRB

+++ Impossible bosses/Assets/Scripts/MainSceneStartUpVars.cs
@@ -7,7 +7,7 @@
     public static int Ehp = 10;
     public static int Emana = 10;
     public static float Emovesp = 10;
-    public static float[] spells = new float[6];
+    public static int[] spells = new int[6];
 
 }
